Modify regex so nerdtree-git-plugin markings are not highlighted.
The nerdtree-git-plugin plugin decorates the filenames in NERDTree with
glyphs indicating the files' status changes. These glyphs are
highlighted differently than the filename. However, when this plugin
also is used, it overrides the glyph's highlighting with its syntax
match definitions.

This commit modifies the regexes on the syn match statements so that any
leading characters in square brackets are ignored, and only the filename
is highlighted.